+++ src/index.tsx
@@ -2,6 +2,7 @@ import { Alert } from './Alert';
 import { Button } from './Button';
 import { Card } from './Card';
 import { Box, Flex, Grid, GlobalStyles } from './Core';
+import { FormField } from './FormField';
 import { Input } from './Input';
 import { Label } from './Label';
 import { Modal } from './Modal';
@@ -18,6 +19,7 @@ export {
   Box,
   Card,
   Flex,
+  FormField,
   Grid,
   GlobalStyles,
   Input,