+++ src/theme/light.tsx
@@ -1,11 +1,11 @@
 import shared from './shared';
 
 const colors = {
-  primary: '#f5932f',
+  primary: '#2378e1',
   secondary: '#2378e1',
   light: '#f3f1f5',
   greyLight: '#a4a1b1',
-  grey: '#686575',
+  grey: '#A4A1B1',
   greyDark: '#403d4d',
   dark: '#0e0b11',
   confirm: '#49ba69',
@@ -18,6 +18,7 @@ const colors = {
   text: '#111213',
   textGrey: '#111213',
   textFade: '#b3b7bc',
+  //
 };
 export default {
   ...shared,
@@ -129,12 +130,12 @@ export default {
     },
     heading2: {
       fontWeight: 400,
-      color: colors.secondary,
+      color: 'primary',
       fontSize: '1.8rem',
     },
     heading3: {
       fontWeight: 600,
-      color: colors.primary,
+      color: 'veriphiPrimary',
       fontSize: '1.4rem',
     },
     cardHeading: {