+++ CHANGELOG.md
@@ -9,9 +9,18 @@ Refer to the [CONTRIBUTING guide](https://github.com/lightspeed/flame/blob/maste
 
 ## [Unreleased]
 
+### Added
+
 - Basic Tooltip component
 - Modal: New aside prop and animations
 - Ultra basic TextLink component
+- FormField Component
+- InputWithAddon Component (use at own risk and peril)
+
+### Breaking
+
+- Input will now always take full-width
+- All components will now use border-box
 
 ## 0.2.0 - 2020-02-15